Removed some casting that is no longer necessary

+++ ecmd/ecmdSpyUser.C
@@ -714,7 +714,7 @@ uint32_t ecmdPutSpyUser(int argc, char * argv[]) {
 
         rc = getSpy(cuTarget, spyName.c_str(), spyBuffer);
 
-        if ((rc == ECMD_SPY_GROUP_MISMATCH) && (numBits == (uint32_t) spyData.bitLength)) {
+        if ((rc == ECMD_SPY_GROUP_MISMATCH) && (numBits == spyData.bitLength)) {
           /* We will go on if the user was going to write the whole spy anyway */
           ecmdOutputWarning("putspy - Problems reading group spy - found a mismatch - going ahead with write\n");
           rc = 0;
